// MarketWatch-Frontend/src/app/models/portfolio.ts
export class Portfolio{
    portfolio_id: number;
    user_id: number;
    stock_id: number;
    quantity: number;
    investedPrice: number;
    portfolio_type: string;
    esg_score: number;
    live_price: number;
    gain_loss_percent: number;
    amount: number;
    constructor(){
       
    }
}